[NFC][lldb][Windows] extract IOHandlerProcessSTDIOWindows
